+++ src/backend/resolvers/mutations/line/index.js
@@ -1,0 +1,9 @@
+const createLine = require('./create')
+const updateLine = require('./update')
+const deleteLine = require('./delete')
+
+module.exports = {
+  createLine,
+  updateLine,
+  deleteLine,
+}