+++ src/backend/resolvers/mutations/line/update.js
@@ -1,0 +1,20 @@
+const { ObjectId } = require('mongodb')
+
+const updateLine = (
+  parent,
+  { input: { _id, name } },
+  { pulseCoreDb },
+  info
+) => pulseCoreDb.collection('lines')
+  .findOneAndUpdate(
+    { _id: ObjectId(_id) },
+    {
+      $set: {
+        name
+      }
+    },
+    { returnOriginal: false }
+  )
+  .then(({ value }) => value)
+
+module.exports = updateLine