+++ src/backend/resolvers/queries/index.js
@@ -17,6 +17,7 @@ const businessObjects = require('./businessObjects')
 const workbooks = require('./workbooks')
 const treatmentPlans = require('./treatmentPlans')
 const regionalTargetingData = require('./regionalTargetingData')
+const lines = require('./lines')
 
 const payerProjects = require('./payerProjects')
 
@@ -41,6 +42,7 @@ module.exports = {
   businessObjects,
   workbooks,
   treatmentPlans,
+  lines,
   ...payerProjects,
   regionalTargetingData,
   cMsOrgPrimarySpecialtyCounts,