+++ src/backend/typeDefs/mutations/line/delete.js
@@ -1,0 +1,14 @@
+const { gql } = require('apollo-server-express')
+
+const deleteLineTypeDefs = gql`
+  input DeleteLineInput {
+    _id: ID!
+  }
+
+  type DeleteLinePayload {
+    _id: ID
+    name: String
+  }
+`
+
+module.exports = deleteLineTypeDefs